+++ src/interface.ts
@@ -2,10 +2,11 @@ import * as UserUtils from "@sv443-network/userutils";
 import * as compareVersions from "compare-versions";
 import { mode, branch, host, buildNumber, compressionFormat, scriptInfo, initialParams, sessionStorageAvailable } from "./constants.js";
 import { getDomain, waitVideoElementReady, getResourceUrl, getSessionId, getVideoTime, log, setLocale, getLocale, hasKey, hasKeyFor, t, tp, type TrLocale, info, error, onInteraction, getThumbnailUrl, getBestThumbnailUrl, fetchVideoVotes, setInnerHtml, getCurrentMediaType, tl, tlp, PluginError, formatNumber, reloadTab, getVideoElement, getVideoSelector, getLikeDislikeBtns, fetchITunesAlbumInfo } from "./utils/index.js";
+import { MultiNanoEmitter } from "./utils/MultiNanoEmitter.js";
 import { addSelectorListener } from "./observers.js";
 import { defaultData, getFeatures, setFeatures } from "./config.js";
 import { autoLikeStore, featInfo, fetchLyricsUrlTop, getLyricsCacheEntry, isIgnoredInputElement, sanitizeArtists, sanitizeSong } from "./features/index.js";
-import { allSiteEvents, type SiteEventsMap } from "./siteEvents.js";
+import { allSiteEvents, siteEvents, type SiteEventsMap } from "./siteEvents.js";
 import { type FeatureConfig, type FeatureInfo, type LyricsCacheEntry, type PluginDef, type PluginInfo, type PluginRegisterResult, type PluginDefResolvable, type PluginEventMap, type PluginItem, type BytmObject, type AutoLikeData, type InterfaceFunctions } from "./types.js";
 import { showPrompt } from "./dialogs/prompt.js";
 import { BytmDialog } from "./components/BytmDialog.js";
@@ -133,6 +134,12 @@ const globalFuncs: InterfaceFunctions = purifyObj({
   getLikeDislikeBtns,
   isIgnoredInputElement,
 
+  // site events:
+  onSiteEvent: siteEvents.on,
+  onceSiteEvent: siteEvents.once,
+  onMultiSiteEvents: siteEvents.onMulti,
+  onceMultiSiteEvents: siteEvents.onceMulti,
+
   // translations:
   /*🔒*/ setLocale: setLocaleInterface,
   getLocale,
@@ -170,6 +177,7 @@ const globalFuncs: InterfaceFunctions = purifyObj({
 
   // other:
   formatNumber,
+  MultiNanoEmitter,
 });
 
 /** Initializes the BYTM interface */